backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
stripe
supabase
psycopg2-binary
pydantic 